Filtros para erros de digitação
Adicionado modificações e funções para contorna pequenos erros de digitação na calculadora

+++ src/Calculadora/operações/fatorial.java
@@ -7,12 +7,13 @@ package Calculadora.operações;
 
 import Calculadora.IOperações;
 import Calculadora.Operações;
+import java.text.DecimalFormat;
 
 /**
  *
  * @author Fabricio Junior
  */
-public class fatorial extends Operações implements IOperações{
+public class fatorial extends Operações implements IOperações {
 
     public fatorial() {
         binaria = false;
@@ -22,7 +23,7 @@ public class fatorial extends Operações implements IOperações{
     @Override
     public String operação(String num1) {
         long retorno = 1;
-        int n1 = Integer.parseInt(num1);
+        int n1 = (int) Float.parseFloat(num1);
         for (int x = 1; x <= n1; x++) {
             retorno = retorno * x;
         }
@@ -33,5 +34,5 @@ public class fatorial extends Operações implements IOperações{
     public String operação(String num1, String num2) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
-     
+
 }
